MRI/requirements.txt
# Web框架和工具
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
websockets==11.0.3
pydantic==2.3.0

# 数据处理和科学计算
numpy==1.24.3
scipy==1.11.2
pandas==2.1.0
h5py==3.9.0

# 图像处理
Pillow==10.0.0
opencv-python==4.8.0.76
scikit-image==0.21.0

# 深度学习
torch==2.0.1
torchvision==0.15.2

# 数据可视化
matplotlib==3.7.2

# 工具
pyyaml==6.0.1
python-dotenv==1.0.0

# 测试
pytest==7.4.0
httpx==0.24.1 